Refactor graphQL request for auth strategies.
