+++ frontend/lib/auth.ts
@@ -1,4 +1,5 @@
 export interface User {
+  id: string;
   authId: string;
   name: string;
   emails: string[];
