+++ frontend/lib/server/graphql.js
@@ -6,30 +6,22 @@ const client = createClient({
 });
 
 const getOrCreateUser = async ({ authId, name }) => {
-  try {
-    const reqBody = gql`
-      mutation($authId: ID!, $name: String!) {
-        getOrCreateUser(newUser: { authId: $authId, name: $name }) {
-          id
-          name
-          authId
-          isPlatformAdmin
-        }
+  const query = gql`
+    mutation($authId: ID!, $name: String!) {
+      getOrCreateUser(newUser: { authId: $authId, name: $name }) {
+        id
+        name
+        authId
+        isPlatformAdmin
       }
-    `;
+    }
+  `;
 
-    return await client
-      .mutation(reqBody, { authId, name })
-      .toPromise()
-      .then((result) => {
-        if (result.error) {
-          throw result.error;
-        } else {
-          return result.data;
-        }
-      });
-  } catch (err) {
-    console.log("getOrCreateUser Error: ", err);
+  const result = await client.mutation(query, { authId, name }).toPromise();
+  if (result.error) {
+    throw result.error;
+  } else {
+    return result.data;
   }
 };
