+++ frontend/tsconfig.json
@@ -15,5 +15,5 @@
     "jsx": "preserve"
   },
   "exclude": ["node_modules"],
-  "include": ["next-env.d.ts", "**/*.ts", "**/*.tsx", "jest.setup.js"]
+  "include": ["next-env.d.ts", "**/*.ts", "**/*.tsx", "jest.setup.js", "lib/server/*.js"]
 }
